updated regex to extract keywords, added CreateRecipeViewModel
